(fix): refactoring endpoints, service and repo

+++ BusinessLayer/Services/BookService.cs
@@ -6,33 +6,51 @@ namespace BusinessLayer.Services
     public class BookService : IBookService
     {
         private readonly IBookRepository _bookRepository;
-        public BookService(IBookRepository repo) => _bookRepository = repo;
 
-        //GET
-        public async Task<IEnumerable<Book>> GetAll() 
-        { 
+        public BookService(IBookRepository repo)
+        {
+            _bookRepository = repo;
+        }
+
+        // GET all books
+        public async Task<IEnumerable<Book>> GetAll()
+        {
             return await _bookRepository.GetAll();
         }
-        public async Task<Book?> GetById(int id) 
-        { 
+
+        // GET book by id
+        public async Task<Book?> GetById(int id)
+        {
+            if (id <= 0) 
+                return null;
+
             return await _bookRepository.GetById(id);
         }
 
-        //POST
-        public async Task<Book> PostBook(Book book) 
-        { 
+        // POST new book
+        public async Task<Book> PostBook(Book book)
+        {
+            if (book == null) 
+                throw new ArgumentNullException(nameof(book));
+
             return await _bookRepository.PostBook(book);
         }
 
-        //PUT
-        public async Task<Book> PutBook(int id, Book book) 
-        { 
+        // PUT update book
+        public async Task<Book?> PutBook(int id, Book book)
+        {
+            if (book == null || id != book.Id) 
+                return null;
+
             return await _bookRepository.PutBook(id, book);
         }
 
-        //DELETE
-        public async Task<Book?> DeleteBook(int id) 
-        { 
+        // DELETE book
+        public async Task<Book?> DeleteBook(int id)
+        {
+            if (id <= 0) 
+                return null;
+
             return await _bookRepository.DeleteBook(id);
         }
     }
